Refactor VertexBuffer struct, move to its own file
