+++ src/render/mod.rs
@@ -3,4 +3,5 @@ pub mod traits;
 pub mod program;
 pub mod shader;
 pub mod vertex;
+pub mod vertex_buffer;
 pub mod window;
